requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-configurations==2.4
django-filter==22.1
django-jazzmin==2.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filters==1.3.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pep8==1.7.1
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
regex==2022.10.31
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
